+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/data/Caption.kt
@@ -12,5 +12,5 @@ data class Caption(
     val isPartial: Boolean,
     val content: String,
     val items: Array<TranscriptItem>?,
-    val entityContentSet: MutableSet<String>?
+    val entities: MutableSet<String>?
 )